pyproject.toml
[project]
name = "nonebot-plugin-shutdown-hook"
version = "0.1.0"
description = "高优先级关闭信号钩子插件"
authors = [
    {name = "Sclock",email = "1342810270@qq.com"},
]
dependencies = [
    "nonebot2>=2.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["src"]

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.13.2",
    "black>=24.4.2",
    "ruff>=0.4.3",
    "pre-commit>=3.7.0",
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot2[fastapi]>=2.3.0",
]

[tool.pdm.scripts]
format = { composite = ["isort ./src/ ","black ./src/ ","ruff check ./src/"] }


[tool.black]
line-length = 110
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 110
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 110
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

